Move StyleConfiguration to non-internal package and export
