+++ tests/org.corpus_tools.hexatomic.grid.tests/src/main/java/org/corpus_tools/hexatomic/grid/internal/style/TestStyleConfiguration.java
@@ -2,6 +2,7 @@ package org.corpus_tools.hexatomic.grid.internal.style;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+import org.corpus_tools.hexatomic.grid.style.StyleConfiguration;
 import org.eclipse.nebula.widgets.nattable.config.CellConfigAttributes;
 import org.eclipse.nebula.widgets.nattable.config.ConfigRegistry;
 import org.eclipse.nebula.widgets.nattable.style.CellStyleAttributes;
